# backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Бэкенд «Умного города» — обращения жителей МКД к управляющей компании, плановые работы, справочники домов/УК/аварийных служб. Один инстанс = одна УК.\n\nРоли: representative (представитель УК), dispatcher (диспетчер), resident (житель). Роль определяется самим пользователем на бэкенде — в запросах отдельно не передаётся.",
        "title": "MAXITO API",
        "contact": {
            "name": "MAXITO"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/dispatcher/company": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Одна запись на весь инстанс, не привязана к дому. Если Представитель ещё не заполнил — пустой объект \"{}\", не ошибка.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "company-emergency"
                ],
                "summary": "Реквизиты управляющей компании",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/maxito_internal_models.Company"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/dispatcher/emergency-services": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Общий на всю систему список, не привязан к дому.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "company-emergency"
                ],
                "summary": "Список аварийных служб",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/maxito_internal_models.EmergencyService"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/representative/company": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Одна запись на весь инстанс, не привязана к дому. Если Представитель ещё не заполнил — пустой объект \"{}\", не ошибка.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "company-emergency"
                ],
                "summary": "Реквизиты управляющей компании",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/maxito_internal_models.Company"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "representative-company"
                ],
                "summary": "Задать/поправить реквизиты УК",
                "operationId": "representativeUpdateCompany",
                "parameters": [
                    {
                        "description": "Реквизиты; full_name обязателен, остальное — нет",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers_representative.UpdateCompanyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/maxito_internal_models.Company"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/representative/emergency-services": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Общий на всю систему список, не привязан к дому.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "company-emergency"
                ],
                "summary": "Список аварийных служб",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/maxito_internal_models.EmergencyService"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/resident/company": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Одна запись на весь инстанс, не привязана к дому. Если Представитель ещё не заполнил — пустой объект \"{}\", не ошибка.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "company-emergency"
                ],
                "summary": "Реквизиты управляющей компании",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/maxito_internal_models.Company"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/resident/emergency-services": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Общий на всю систему список, не привязан к дому.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "company-emergency"
                ],
                "summary": "Список аварийных служб",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/maxito_internal_models.EmergencyService"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apidoc.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code — машиночитаемый код ошибки; присутствует только у части\nэндпоинтов (см. описание конкретного эндпоинта), у остальных отсутствует.",
                    "type": "string",
                    "example": "bad_request"
                },
                "error": {
                    "type": "string",
                    "example": "invalid request"
                }
            }
        },
...
        "internal_handlers_representative.UpdateCompanyRequest": {
            "type": "object",
            "required": [
                "full_name"
            ],
            "properties": {
                "contact_phone": {
                    "type": "string"
                },
                "dispatcher_phone": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "short_name": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                }
            }
        },
...
        "maxito_internal_models.Company": {
            "type": "object",
            "properties": {
                "contact_phone": {
                    "type": "string"
                },
                "dispatcher_phone": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "short_name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                }
            }
        },
...
        "maxito_internal_models.EmergencyService": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT, полученный через POST /auth/max. Передавать как \"Bearer \u0026lt;token\u0026gt;\". Для локальной отладки без MAX бэкенд также принимает X-Max-User-Id (+X-Max-User-Phone при первой привязке), если на сервере включён ALLOW_DEV_HEADERS.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "InternalKey": {
            "description": "Общий секрет между сервером бота и бэкендом — только для /internal/*.",
            "type": "apiKey",
            "name": "X-Internal-Key",
            "in": "header"
        }
    }
}